#66 Replace ThreadDao with ProductDao.threadUpdateProduct() method.
