--- src/main/java/team/asd/dao/ThreadUpdateDao.java
@@ -1,8 +1,0 @@
-package team.asd.dao;
-
-import team.asd.entity.Product;
-
-public interface ThreadUpdateDao {
-
-	void threadUpdateProduct(Product product);
-}
